[modify Code]

+++ ZeroDemo/src/demo/ZeroSum.java
@@ -11,8 +11,8 @@ public class ZeroSum {
 	public static int solution(int []A){
 		int result = -1;
 		int length = A.length;
-		int temp[] = new int[length];
-		int sum = 0;	
+		double temp[] = new double[length];
+		double sum = 0;	
 		for(int i =0; i < length; i++){
 			sum += A[i];
 			temp[i] = sum;
